+++ lib/pbio/src/port_lump.c
@@ -31,16 +31,15 @@
 
 #define EV3_UART_MAX_MESSAGE_SIZE   (LUMP_MAX_MSG_SIZE + 3)
 
-#define EV3_UART_MAX_DATA_ERR       6
-
 #define EV3_UART_TYPE_MIN           29      // EV3 color sensor
 #define EV3_UART_TYPE_MAX           101
 #define EV3_UART_SPEED_MIN          2400
 #define EV3_UART_SPEED_LPF2         115200  // standard baud rate for Powered Up
 #define EV3_UART_SPEED_MAX          460800  // in practice 115200 is max
 
-#define EV3_UART_DATA_KEEP_ALIVE_TIMEOUT    100 /* msec */
-#define EV3_UART_IO_TIMEOUT                 250 /* msec */
+#define EV3_UART_DATA_KEEP_ALIVE_TIMEOUT    100 // msec
+#define EV3_UART_DATA_KEEP_ALIVE_MAX_MISSED 2   // Allow almost (1 + 2) * 100ms of no data
+#define EV3_UART_IO_TIMEOUT                 250 // msec
 
 enum ev3_uart_info_bit {
     EV3_UART_INFO_BIT_CMD_TYPE,
@@ -173,7 +172,7 @@ struct _pbio_port_lump_dev_t {
     uint32_t tx_msg_size;
     /** Size of the current message being received. */
     uint32_t rx_msg_size;
-    /** Total number of errors that have occurred. */
+    /** Total number of errors that have occurred. Re-used in different stages of synchronization and data reading. */
     uint32_t err_count;
     /** Flag that indicates that good DATA lump_dev->msg has been received since last watchdog timeout. */
     bool data_rec;
@@ -921,7 +920,9 @@ pbio_error_t pbio_port_lump_data_send_thread(pbio_os_state_t *state, pbio_port_l
     PBIO_OS_ASYNC_BEGIN(state);
 
     // Some devices need the NACK keep-alive signal before doing anything, so
-    // initially set the timer to expire soon.
+    // initially set the timer to expire soon. Pretend we have data so this
+    // first timer event does not count as an error.
+    lump_dev->data_rec = true;
     pbio_os_timer_set(timer, 1);
 
     for (;;) {
@@ -930,13 +931,23 @@ pbio_error_t pbio_port_lump_data_send_thread(pbio_os_state_t *state, pbio_port_l
 
         // Handle keep alive timeout
         if (pbio_os_timer_is_expired(timer)) {
-            // Make sure we are receiving data. The first time around, we allow
-            // not having any data yet.
-            if (!lump_dev->data_rec && timer->duration == EV3_UART_DATA_KEEP_ALIVE_TIMEOUT) {
-                debug_pr("No data since last keepalive\n");
+            // Make sure we are receiving data. Initialized as true above, so
+            // the first pass always clears the error count.
+            if (lump_dev->data_rec) {
+                lump_dev->err_count = 0;
+            } else {
+                lump_dev->err_count++;
+                debug_pr("No data since last keepalive. Count: %d\n", lump_dev->err_count);
+            }
+
+            // Bail out if we don't receive data anymore. Allow a few missed
+            // samples since some devices use our keep alive heartbeat to send
+            // idle messages, meaning they'll get sent just after one timeout.
+            if (lump_dev->err_count > EV3_UART_DATA_KEEP_ALIVE_MAX_MISSED) {
                 lump_dev->status = PBDRV_LEGODEV_LUMP_STATUS_ERR;
                 return PBIO_ERROR_TIMEDOUT;
             }
+
             lump_dev->data_rec = false;
             lump_dev->tx_msg[0] = LUMP_SYS_NACK;
             lump_dev->tx_msg_size = 1;